bugs bugs bugs
Fixed a bug on the Browse page, where if there were no results and the user clicked 'Next'. The API would break.

Fixed an issue with the API that it would display 'Xbox Game Pass' as where to buy 'World of Warcraft' when WoW is exclusively sold on Battle.net